Multiple Indicator Cluster Survey (MICS) 2018 Household Questionnaire Data for Punjab, Pakistan

PakPMICS2018hh provides data set and function for exploration of Multiple Indicator Cluster Survey (MICS) 2017-18 Household questionnaire data for Punjab, Pakistan. The results of the present survey are critically important for the purposes of SDG monitoring, as the survey produces information on 32 global SDG indicators. The data was collected from 53,840 households selected at the second stage with systematic random sampling out of a sample of 2,692 clusters selected using Probability Proportional to size sampling. Six questionnaires were used in the survey: 1. a household questionnaire to collect basic demographic information on all de jure household members (usual residents), the household, and the dwelling; 2. a water quality testing questionnaire administered in three households in each cluster of the sample; 3. a questionnaire for individual women administered in each household to all women age 15-49 years; 4. a questionnaire for individual men administered in every second household to all men age 15-49 years; 5. an under-5 questionnaire, administered to mothers (or caretakers) of all children under 5 living in the household; and 6. a questionnaire for children age 5-17 years, administered to the mother (or caretaker) of one randomly selected child age 5-17 years living in the household.

Format

A data frame with 53840 rows and 330 variables:

HH1

Cluster number

HH2

Household number

HH3

Interviewer number

HH4

Supervisor number

HH5D

Day of interview

HH5M

Month of interview

HH5Y

Year of interview

HH6

Area

HH8

HH selected for Questionnaire for Men

HH9

HH selected for Water Quality Testing

HH10

HH selected for blank testing

HH11H

Start of interview - Hour

HH11M

Start of interview - Minutes

HH12

Consent

HH46

Result of HH interview

HH13H

End of interview - Hour

HH13M

End of interview - Minutes

HH14

Language of the questionnaire

HH15

Language of the interview

HH16

Native language of the respondent

HH17

Translator used

HH26A

Rank number of the selected child

HH26B

Child line number

HH26C

Child's age

HH33

Consent for interview girls 15-17

HH33.01

Roster to check the consent for interview girls 15-17

HH33.02

Roster to check the consent for interview girls 15-17

HH33.03

Roster to check the consent for interview girls 15-17

HH33.04

Roster to check the consent for interview girls 15-17

HH33.05

Roster to check the consent for interview girls 15-17

HH33.06

Roster to check the consent for interview girls 15-17

HH33.07

Roster to check the consent for interview girls 15-17

HH33.08

Roster to check the consent for interview girls 15-17

HH33.09

Roster to check the consent for interview girls 15-17

HH33.10

Roster to check the consent for interview girls 15-17

HH33.11

Roster to check the consent for interview girls 15-17

HH33.12

Roster to check the consent for interview girls 15-17

HH33.13

Roster to check the consent for interview girls 15-17

HH33.14

Roster to check the consent for interview girls 15-17

HH33.15

Roster to check the consent for interview girls 15-17

HH33.16

Roster to check the consent for interview girls 15-17

HH33.17

Roster to check the consent for interview girls 15-17

HH33.18

Roster to check the consent for interview girls 15-17

HH33.19

Roster to check the consent for interview girls 15-17

HH33.20

Roster to check the consent for interview girls 15-17

HH33.21

Roster to check the consent for interview girls 15-17

HH33.22

Roster to check the consent for interview girls 15-17

HH33.23

Roster to check the consent for interview girls 15-17

HH33.24

Roster to check the consent for interview girls 15-17

HH33.25

Roster to check the consent for interview girls 15-17

HH33.26

Roster to check the consent for interview girls 15-17

HH33.27

Roster to check the consent for interview girls 15-17

HH33.28

Roster to check the consent for interview girls 15-17

HH33.29

Roster to check the consent for interview girls 15-17

HH33.30

Roster to check the consent for interview girls 15-17

HH33.31

Roster to check the consent for interview girls 15-17

HH33.32

Roster to check the consent for interview girls 15-17

HH33.33

Roster to check the consent for interview girls 15-17

HH33.34

Roster to check the consent for interview girls 15-17

HH33.35

Roster to check the consent for interview girls 15-17

HH33.36

Roster to check the consent for interview girls 15-17

HH33.37

Roster to check the consent for interview girls 15-17

HH33.38

Roster to check the consent for interview girls 15-17

HH33.39

Roster to check the consent for interview girls 15-17

HH33.40

Roster to check the consent for interview girls 15-17

HH33.41

Roster to check the consent for interview girls 15-17

HH33.42

Roster to check the consent for interview girls 15-17

HH33.43

Roster to check the consent for interview girls 15-17

HH33.44

Roster to check the consent for interview girls 15-17

HH33.45

Roster to check the consent for interview girls 15-17

HH33.46

Roster to check the consent for interview girls 15-17

HH33.47

Roster to check the consent for interview girls 15-17

HH33.48

Roster to check the consent for interview girls 15-17

HH33.49

Roster to check the consent for interview girls 15-17

HH33.50

Roster to check the consent for interview girls 15-17

HH39

Consent for interview boys 15-17

HH39.01

Roster to check the consent for interview boys 15-17

HH39.02

Roster to check the consent for interview boys 15-17

HH39.03

Roster to check the consent for interview boys 15-17

HH39.04

Roster to check the consent for interview boys 15-17

HH39.05

Roster to check the consent for interview boys 15-17

HH39.06

Roster to check the consent for interview boys 15-17

HH39.07

Roster to check the consent for interview boys 15-17

HH39.08

Roster to check the consent for interview boys 15-17

HH39.09

Roster to check the consent for interview boys 15-17

HH39.10

Roster to check the consent for interview boys 15-17

HH39.11

Roster to check the consent for interview boys 15-17

HH39.12

Roster to check the consent for interview boys 15-17

HH39.13

Roster to check the consent for interview boys 15-17

HH39.14

Roster to check the consent for interview boys 15-17

HH39.15

Roster to check the consent for interview boys 15-17

HH39.16

Roster to check the consent for interview boys 15-17

HH39.17

Roster to check the consent for interview boys 15-17

HH39.18

Roster to check the consent for interview boys 15-17

HH39.19

Roster to check the consent for interview boys 15-17

HH39.20

Roster to check the consent for interview boys 15-17

HH39.21

Roster to check the consent for interview boys 15-17

HH39.22

Roster to check the consent for interview boys 15-17

HH39.23

Roster to check the consent for interview boys 15-17

HH39.24

Roster to check the consent for interview boys 15-17

HH39.25

Roster to check the consent for interview boys 15-17

HH39.26

Roster to check the consent for interview boys 15-17

HH39.27

Roster to check the consent for interview boys 15-17

HH39.28

Roster to check the consent for interview boys 15-17

HH39.29

Roster to check the consent for interview boys 15-17

HH39.30

Roster to check the consent for interview boys 15-17

HH39.31

Roster to check the consent for interview boys 15-17

HH39.32

Roster to check the consent for interview boys 15-17

HH39.33

Roster to check the consent for interview boys 15-17

HH39.34

Roster to check the consent for interview boys 15-17

HH39.35

Roster to check the consent for interview boys 15-17

HH39.36

Roster to check the consent for interview boys 15-17

HH39.37

Roster to check the consent for interview boys 15-17

HH39.38

Roster to check the consent for interview boys 15-17

HH39.39

Roster to check the consent for interview boys 15-17

HH39.40

Roster to check the consent for interview boys 15-17

HH39.41

Roster to check the consent for interview boys 15-17

HH39.42

Roster to check the consent for interview boys 15-17

HH39.43

Roster to check the consent for interview boys 15-17

HH39.44

Roster to check the consent for interview boys 15-17

HH39.45

Roster to check the consent for interview boys 15-17

HH39.46

Roster to check the consent for interview boys 15-17

HH39.47

Roster to check the consent for interview boys 15-17

HH39.48

Roster to check the consent for interview boys 15-17

HH39.49

Roster to check the consent for interview boys 15-17

HH39.50

Roster to check the consent for interview boys 15-17

HH44

Consent for Water Quality Testing

HH47

Respondent to HH questionnaire

HH48

Number of HH members

HH49

Number of women 15 - 49 years

HH50

Number of men 15 - 49 years

HH51

Number of children under age 5

HH52

Number of children age 5-17

INTROHL

Introduction HH Listing

HHAUX

Other household memebers

HHWQ

Water Quality prompt

HHFIN

Finish

HC1B

Language of household head

HC3

Number of rooms used for sleeping

HC4

Main material of floor

HC5

Main material of roof

HC6

Main material of exterior wall

HC7A

Household have: Fixed telephone line

HC7B

Household have: Radio

HC7C

Household have: Gas Heater

HC7D

Household have: Cooking Range

HC7E

Household have: Sewing Machine (without electric motor)

HC7F

Household have: An iron (Gas/ Coal)

HC7G

Household have: Bed

HC7H

Household have: Sofa

HC7I

Household have: Cupboard

HC7J

Household have: Wall Clock

HC8

Household have electricity

HC9A

Household have: Television

HC9B

Household have: Refrigerator

HC9C

Household have: Washing Machine/ Dryer

HC9D

Household have: Air Cooler/ Fan

HC9E

Household have: Microwave Oven

HC9F

Household have: Electric Iron

HC9G

Household have: Water Filter

HC9H

Household have: Donkey Pump/ Turbine

HC9I

Household have: Air conditioner

HC9J

Household have: Sewing Machine (with electirc motor)

HC10A

Any member of household own: Watch

HC10B

Any member of household own: Bicycle

HC10C

Any member of household own: Motorcycle or scooter

HC10D

Any member of household own: Animal-drawn cart

HC10E

Any member of household own: Car, truck or van

HC10F

Any member of household own: Boat with motor

HC10G

Any member of household own: Tractor Trolley

HC10H

Any member of household own: Autorickshaw/ Chingchi

HC11

Any member have a computer

HC12

Any member have a mobile telephone

HC13

Internet access at home

HC14

Household owns the dwelling

HC15

Any household member own land that can be used for agriculture

HC16

Hectares of agricultural land members of household owns

HC17

Household own any animals

HC18A

Milk cows, buffaloes or bulls

HC18B

Other cattle

HC18C

Horses, donkeys, camel or mules

HC18D

Goats

HC18E

Sheep

HC18F

Chickens

HC18H

Ducks/Turkey

HC19

Any household member own bank account

ST2.1

Aware of (name of programme)

ST2.2

Aware of (name of programme)

ST2.3

Aware of (name of programme)

ST2.4

Aware of (name of programme)

ST2.5

Aware of (name of programme)

ST2.6

Aware of (name of programme)

ST3.1

Received assistance through (name of programme)

ST3.2

Received assistance through (name of programme)

ST3.3

Received assistance through (name of programme)

ST3.4

Received assistance through (name of programme)

ST3.5

Received assistance through (name of programme)

ST3.6

Received assistance through (name of programme)

ST4U.1

Last time assistance was received (unit)

ST4U.2

Last time assistance was received (unit)

ST4U.3

Last time assistance was received (unit)

ST4U.4

Last time assistance was received (unit)

ST4U.5

Last time assistance was received (unit)

ST4U.6

Last time assistance was received (unit)

ST4N.1

Last time assistance was received (number)

ST4N.2

Last time assistance was received (number)

ST4N.3

Last time assistance was received (number)

ST4N.4

Last time assistance was received (number)

ST4N.5

Last time assistance was received (number)

ST4N.6

Last time assistance was received (number)

RM1

Working outside this country

RM2

Number of persons working outside this country

RM3

Household receive remittances from outside the country

RM4

Household receive any cash donation from within the country

EU1

Type of cookstove mainly used for cooking

EU2

Cookstove have a chimney

EU3

Cookstove have a fan

EU4

Type of energy source for cookstove

EU5

Place for cooking

EU6

Type of space heating in household

EU7

Space heater have a chimney

EU8

Type of energy source for heater

EU9

Type of lighting in household

WS1

Main source of drinking water

WS2

Main source of water used for other purposes (if bottled water used for drinking)

WS3

Location of the water source

WS4

Time (in minutes) to get water and come back

WS5

Person collecting water

WS6

Number of times person collected water in the last seven days

WS7

There been any time in the last month without sufficient water

WS8

Main reason without sufficient water

WS9

Treat water to make safer for drinking

WS10A

Water treatment: Boil

WS10B

Water treatment: Add bleach/chlorine

WS10C

Water treatment: Strain it through a cloth

WS10D

Water treatment: Use water filter

WS10E

Water treatment: Solar disinfection

WS10F

Water treatment: Let it stand and settle

WS10X

Water treatment: Other

WS10Z

Water treatment: DK

WS10NR

Water treatment: Missing

WS11

Type of toilet facility

WS12

Pit latrine or septic tank ever been emptied

WS13

Place the contents were emptied

WS14

Location of the toilet faciltity

WS15

Toilet facility shared

WS16

Toilet shared with other household or with general public

WS17

Households using this toilet facility

HW1

Place where household members most often wash their hands

HW2

Water available at the place for handwashing

HW3

Soap or detergent present at place of handwashing

HW4

Usual place for handwashing

HW5

Soap or detergent available for washing hands

HW6

Hand washing material shown

HW7A

Bar soap

HW7B

Detergent (Powder / Liquid / Paste)

HW7NR

Missing

SA1

Salt iodization test outcome

SA2

Salt iodization recheck test outcome

WQ1

Cluster number

WQ2

Household number

WQ3

Measurer's number

WQ4

Supervisor's number

WQ5D

Day of interview

WQ5M

Month of interview

WQ5Y

Year of interview

WQ6

HH selected for blank testing

WQ8

Consent for water testing

WQ31

Result of water quality testing

WQ10H

Start of water quality test - Hour

WQ10M

Start of water quality test - Minutes

WQ22H

End of water quality test - Hour

WQ22M

End of water quality test - Minutes

WQ11

Glass of the water that members of your household usually drink

WQ12

whether the water was collected directly from the source or from a separate storage container

WQ13

Have you labled sample

WQ14

Done anything to the water to make it safer to drink

WQ15A

Water treatment: Boil

WQ15B

Water treatment: Add bleach/chlorine

WQ15C

Water treatment: Strain it through a cloth

WQ15D

Water treatment: Use water filter

WQ15E

Water treatment: Solar disinfection

WQ15F

Water treatment: Let it stand and settle

WQ15X

Water treatment: Other

WQ15Z

Water treatment: DK

WQ15NR

Water treatment: Missing

WQ16

Was the glass water from the main source

WQ17

Source of water where the glass of water was collected

WQ18

Main source of water was shown

WQ19

Source water sample collected

WQ20

HH selected for blank testing

WQ21

Blank water test available

WQ23H

End of water quality test - Hour

WQ23M

End of water quality test - Minutes

WQ24D

Day of of recording test results

WQ24M

Month of of recording test results

WQ24Y

Year of of recording test results

WQ26

Household water test (100ml)

WQ27

Source water test (100ml)

WQ29

Blank water test (100ml)

WQ25M

End of water quality test - Hour

WQ25H

End of water quality test - Minutes

HHSEX

Sex of household head

HHAGE

Age of household head

stratum

stratum

psu

psu

wscore

Combined wealth score

windex5

Wealth index quintile

windex10

Percentile Group of com1

wscoreu

Urban wealth score

windex5u

Urban wealth index quintile

windex10u

Percentile Group of urb1

wscorer

Rural wealth score

windex5r

Rural wealth index quintile

windex10r

Percentile Group of rur1

division

Division

majorcity

majorcity

hh6r

Area

suburban

hh7

Districts

helevel

Education of household head

HH53

Number of woman' questionnaires completed

HH54

Number of man' questionnaires completed

HH50A

Number of men 15 - 49 years in interviewed households

HH55

Number of under - 5 questionnaires completed

HH56

Number of CF questionnaires completed (age 5-17 years)

hhweight

Household sample weight

wqhweight

Weight for household WQT

wshweight

Weight for source WQT

wqsweight

Weight for source WQT

References

UNICEF MICS.

Author

Muhammad Yaseen (myaseen208@gmail.com)

Examples

# Loading PakPMICS2018hh
 if (FALSE) { # \dontrun{
 load(url(
  "https://github.com/myaseen208/PakPMICS2018Data/raw/master/PakPMICS2018hh.RData"
  ))
  PakPMICS2018hh
 } # }