Multiple Indicator Cluster Survey (MICS) 2018 Maternal Mortality Data for Punjab, Pakistan

PakPMICS2018mm data of Maternal Mortality, MICS Punjab, 2014.

Format

A data frame with 157899 rows and 62 variables:

HH1

Cluster number

HH2

Household number

WM1

Cluster number

WM2

Household number

WMINT

Interviewer number

WM3

Woman's line number

LN

Line number

MMLN

Sibling's line number

MM15

Sibling's gender

MM16

Sibling still alive

MM17

Age of sibling

MM18

Years since death

MM19

Age at death of sibling

MM22

Pregnant when died

MM23

Died during the childbirth

MM24

Died within two months after the end of a pregnancy or childbirth

MM25

Days after the end of the pregnancy or childbirth did die

MM26

Death due to an act of violence

MM27

Death due to an accident

HH4

Supervisor number

HH6

Area

WDOI

Date of interview women (CMC)

WDOB

Date of birth of woman (CMC)

wscore

Combined wealth score

windex5

Wealth index quintile

windex10

Percentile Group of com1

wscoreu

Urban wealth score

windex5u

Urban wealth index quintile

windex10u

Percentile Group of urb1

wscorer

Rural wealth score

windex5r

Rural wealth index quintile

windex10r

Percentile Group of rur1

division

Division

majorcity

majorcity

hh6r

Area

suburban

hh7

Districts

welevel

Education

disability

Functional difficulties (age 18-49 years)

insurance

Health insurance

wmweight

Woman's sample weight

stratum

stratum

psu

psu

References

UNICEF MICS.

Author

Muhammad Yaseen (myaseen208@gmail.com)

Examples

# Loading PakPMICS2018mm
 if (FALSE) { # \dontrun{
 load(url(
  "https://github.com/myaseen208/PakPMICS2018Data/raw/master/PakPMICS2018mm.RData"
  ))
  PakPMICS2018mm
 } # }