Multiple Indicator Cluster Survey (MICS) 2018 Men Data for Punjab, Pakistan

PakPMICS2018mn data of Men, MICS Punjab, 2014.

Format

A data frame with 157899 rows and 62 variables:

HH1

Cluster number

HH2

Household number

MWM1

Cluster number

MWM2

Household number

MWMINT

Interviewer number

MWM3

Man's line number

LN

Line number

MWM4

Supervisor's number

MWM5

Interviewer's number

MWM6D

Day of interview

MWM6M

Month of interview

MWM6Y

Year of interview

MWM8

Household interviewer

MWM9

Consent

MWM17

Result of man's interview

MWM7H

Start of interview - Hour

MWM7M

Start of interview - Minutes

MWM10H

End of interview - Hour

MWM10M

End of interview - Minutes

MWM11

Interview completed in private

MWM12

Language of the Questionnaire

MWM13

Language of the Interview

MWM14

Native language of the Respondent

MWM15

Translator used

MWMHINT

Household interviewer

MWB1

Check: Respondent to HH questionnaire

MWB2

Check: Highest level of school attended

MWB3M

Month of birth of man

MWB3Y

Year of birth of man

MWB4

Age of man

MWB5

Ever attended school

MWB6A

Highest level of school attended

MWB6B

Highest grade attended at that level

MWB7

Grade completion

MWB9

Attended school during current school year

MWB10A

Level of education attended current school year

MWB10B

Grade of education attended current school year

MWB11

Attended school previous school year

MWB12A

Level of education attended previous school year

MWB12B

Grade of education attended previous school year

MWB14

Can read part of the sentence

MWB15

Duration of living in current place

MWB16

Place of living prior to moving to current place

MWB17

Region prior to moving to current place

MWB18

Covered by health insurance

MWB19A

Type of health insurance: Public health organization

MWB19B

Type of health insurance: Health insurance through employer

MWB19C

Type of health insurance: Social security

MWB19D

Type of health insurance: Other privately purchased commercial health insurance

MWB19X

Type of health insurance: Other

MWB19NR

Type of health insurance: Missing

MWDOI

Date of interview man (CMC)

MWB1C

Date of birth of man (CMC)

MWB1F

Date flag for WB1

MWB2C

Age of respondent (calculated)

MWAGE

Age

MMT1

Frequency of reading newspaper or magazine

MMT2

Frequency of listening to the radio

MMT3

Frequency of watching TV

MMT4

Ever used a computer or a tablet

MMT5

Computer / tablet usage in the last 3 months

MMT6A

During the last 3 months: Copy or move a file or folder

MMT6B

During the last 3 months: Use a copy / paste in document

MMT6C

During the last 3 months: Send e-mail with attached file

MMT6D

During the last 3 months: Use a basic arithmetic formula in a spreadsheet

MMT6E

During the last 3 months: Connect and install a new device

MMT6F

During the last 3 months: Install and configure software

MMT6G

During the last 3 months: Create an electronic presentation

MMT6H

During the last 3 months: Transfer a file

MMT6I

During the last 3 months: Write a computer program

MMT9

Ever used internet

MMT10

Internet usage in the last 3 months

MMT11

Own a mobile phone

MMT12

Mobile phone usage in the last 3 months

MCM1

Ever fathered any childeren

MCM2

Any sons or daughters living with you

MCM3

Sons living with you

MCM4

Daughters living with you

MCM5

Any sons or daughters not living with you

MCM6

Sons living elsewhere

MCM7

Daughters living elsewhere

MCM8

Ever had child who later died

MCM9

Boys dead

MCM10

Girls dead

MCM11

Children ever born

MCM12

Confirm total number of ever born

MCM15

Same biological mother

MCM16

Number of women respondent fathered children with

MCM17

Age when first child was born

MCM18M

Month of last birth

MCM18Y

Year of last birth

MDV1A

If she goes out with out telling husband: wife beating justified

MDV1B

If she neglects the children: wife beating justified

MDV1C

If she argues with husband: wife beating justified

MDV1D

If she refuses sex with husband: wife beating justified

MDV1E

If she burns the food: wife beating justified

MVT1

Something taken by using force or someone threatening to use force

MVT2

Victimisation happened during the last 12 months

MVT3

Number of times victimisation happened in the last year

MVT5

The last time that this happened, was anything stolen

MVT6

Person(s) had a weapon

MVT7A

Offender(s) had a knife, a gun or something else used as a weapon: A knife

MVT7B

Offender(s) had a knife, a gun or something else used as a weapon: A gun

MVT7X

Offender(s) had a knife, a gun or something else used as a weapon: Other weapon

MVT7NR

Offender(s) had a knife, a gun or something else used as a weapon: Missing

MVT8

Incident reported to the police

MVT9

Physically attacked

MVT10

Attack happened during the last 12 months

MVT11

Number of times attack happened in the last year

MVT12

Place attack happened the last time

MVT13

Number of people involved in committing the offence

MVT14

At least one of the offender(s) recognized at the time of the incident

MVT17

Person(s) had a weapon

MVT18A

Offender(s) had a knife, a gun or something else used as a weapon: A knife

MVT18B

Offender(s) had a knife, a gun or something else used as a weapon: A gun

MVT18X

Offender(s) had a knife, a gun or something else used as a weapon: Other weapon

MVT18NR

Offender(s) had a knife, a gun or something else used as a weapon: Missing

MVT19

Incident reported to the police

MVT20

Feeling safe walking alone in neighbourhood after dark

MVT21

Feeling safe at home alone after dark

MVT22A

In the past 12 months, felt discriminated: Ethnic or immigration origin

MVT22B

In the past 12 months, felt discriminated: Gender

MVT22C

In the past 12 months, felt discriminated: Sexual orientation

MVT22D

In the past 12 months, felt discriminated: Age

MVT22E

In the past 12 months, felt discriminated: Religion or belief

MVT22F

In the past 12 months, felt discriminated: Disability

MVT22X

In the past 12 months, felt discriminated: any other reason

MMA1

Currently married

MMA3

More than one wives

MMA4

Number of wives

MMA5

Ever married

MMA6

Marital status

MMA7

Married with someone once or more than once

MMA8M

Month of first marriage

MMA8Y

Year of first marriage

MMA11

Age at first marriage

MHA1

Ever heard of HIV or AIDS

MHA2

Can avoid HIV by having one uninfected partner

MHA3

Can get HIV from mosquito bites

MHA4

Can avoid HIV by using a condom correctly every time

MHA5

Can get HIV by sharing food with a person who has HIV

MHA6

Can get HIV through supernatural means

MHA7

Healthy-looking person may have HIV

MHA8A

HIV from mother to child during pregnancy

MHA8B

HIV from mother to child during delivery

MHA8C

HIV from mother to child through breastfeeding

MHA10

Aware of drugs used to reduce the risk of transmission to the baby

MHA24

Ever been tested for HIV

MHA25

Most recent time of testing for HIV

MHA26

Received results of HIV test

MHA27

Know a place to get HIV test

MHA28

Heard of test kit for HIV testing

MHA29

Ever tested yourself using test kit

MHA30

Would buy fresh vegetables from shopkeeper with AIDS virus

MHA31

Children living with HIV should be allowed to attend school with other children

MHA32

People hesitate to take an HIV test because they are afraid of how other people will react if the test result is positive

MHA33

People talk badly about people living with HIV, or who are thought to be living with HIV

MHA34

People living with HIV, or thought to be living with HIV, lose the respect of other people

MHA35

Agreement with the following statement: I would be ashamed if someone in my family had HIV

MHA36

Fear of getting HIV in contact with the saliva of a person living with HIV

MTA1

Ever tried cigarette smoking

MTA2

Age when cigarette was smoked for the first time

MTA3

Currently smoking cigarettes

MTA4

Number of cigarettes smoked in the last 24 hours

MTA5

Number of days when cigarettes were smoked in past month

MTA6

Ever tried any smoked tobacco products other than cigarettes

MTA7

Used any smoked tobacco products during the last month

MTA8A

Type of smoked tobacco product: Cigars

MTA8B

Type of smoked tobacco product: Water pipe

MTA8C

Type of smoked tobacco product: Cigarillos

MTA8D

Type of smoked tobacco product: Pipe/SHESHA

MTA8X

Type of smoked tobacco product: Other

MTA8NR

Type of smoked tobacco product: No response

MTA9

Number of days when tobacco products where smoked in past month

MTA10

Ever tried any form of smokeless tobacco products

MTA11

Used any smokeless tobacco products during the last month

MTA12A

Type of smokeless tobacco product used: Chewing tobacco

MTA12B

Type of smokeless tobacco product used: Snuff

MTA12D

Type of smokeless tobacco product used: PAN WITH TOBACCO

MTA12E

Type of smokeless tobacco product used: GUTKA

MTA12F

Type of smokeless tobacco product used: NASWAR

MTA12G

Type of smokeless tobacco product used: MAWA TOBACCO

MTA12H

Type of smokeless tobacco product used: PAN WITH TOBACCO

MTA12X

Type of smokeless tobacco product used: Other

MTA12NR

Type of smokeless tobacco product used: No response

MTA13

Number of days when smokeless tobacco products where used in past month

MLS1

Estimation of overall happiness

MLS2

Satisfaction with ladder step

MLS3

Life satisfaction in comparison with last year

MLS4

Life satisfaction expectation one year from now

HH4

Supervisor number

HH6

Area

MWDOB

Date of birth of man (CMC)

MWDOM

Date of marriage of man (CMC)

MWAGEM

Age at first marriage of man

MMSTATUS

Marital status of man

disability

Functional difficulties (age 18-49 years)

wscore

Combined wealth score

windex5

Wealth index quintile

windex10

Percentile Group of com1

wscoreu

Urban wealth score

windex5u

Urban wealth index quintile

windex10u

Percentile Group of urb1

wscorer

Rural wealth score

windex5r

Rural wealth index quintile

windex10r

Percentile Group of rur1

division

Division

majorcity

majorcity

hh6r

Area

suburban

hh7

Districts

mwelevel

Education

minsurance

Health insurance

mdisability

Functional difficulties (age 18-49 years)

mnweight

Men's sample weight

stratum

stratum

psu

psu

References

UNICEF MICS.

Author

Muhammad Yaseen (myaseen208@gmail.com)

Examples

# Loading PakPMICS2018mn
 if (FALSE) { # \dontrun{
 load(url(
  "https://github.com/myaseen208/PakPMICS2018Data/raw/master/PakPMICS2018mn.RData"
  ))
  PakPMICS2018mn
 } # }