Multiple Indicator Cluster Survey (MICS) 2018 Child Questionnaire Data for Punjab, Pakistan

PakPMICS2018hl data of Child questionnaire, MICS Punjab, 2014.

Format

A data frame with 157899 rows and 62 variables:

HH1

Cluster number

HH2

Household number

HL1

Line number

HL3

Relationship to the head

HL4

Sex

HL5M

Month of birth

HL5Y

Year of birth

HL6

Age

HL7

Member stayed in the household last night

HL7A

Age 10 and above

HL7B

Marital Status

HL8

Line number of woman age 15 - 49

HL9

Line number of man age 15 - 49

HL10

Line number for children age 0-4

HL11

Member age 0-17

HL12

Is natural mother alive

HL13

Does natural mother live in HH

HL14

Natural mother's line number in HH

HL15

Where does natural mother live

HL16

Is natural father alive

HL17

Does natural father live in HH

HL18

Natural father's line number in HH

HL19

Where does natural father live

HL20

Line number of morhter or primary caretaker for children 0-17 years of age

HL33

Consent for interview girls 15-17

HL39

Consent for interview boys 15-17

ED1

Line number

ED2A

Age

ED3

Age above 3

ED4

Ever attended school or any Early Childhood Education programme

ED5A

Highest level of education attended

ED5B

Highest grade attended at that level

ED6

Highest grade completed at that level

ED7

Age 3-24

ED8

Check: Ever attended school, Pre School, Katchi or any Early Childhood Education programme

ED9

Attended school during current school year

ED10A

Level of education attended current school year

ED10B

Grade of education attended current school year

ED10C

Going to any school last week for at least 4-5 hours daily

ED11

Attended public school current school year

ED12

School tuition in the current school year

ED13A

Tuition provided: Govt./public

ED13B

Tuition provided: Religious / faith org.

ED13C

Tuition provided: Private

ED13X

Tuition provided: Other

ED13Z

Tuition provided: Z

ED13NR

Tuition provided: NR

ED14

Material support in the current school year

ED15

Attended school previous school year

ED16A

Level of education attended previous school year

ED16B

Grade of education attended previous school year

DA1

Line number

DA2B

Age

DA3

Age 18 or above

DA4

Wear glasses or contact lenses

DA5

Difficulty seeing, even if wearing glasses

DA6

Difficulty seeing

DA7

Wear a hearing aid

DA8

Difficulty hearing, even if using a hearing aid

DA9

Difficulty hearing

DA10

Difficulty communicating when using his/her usual language

DA11

Difficulty remembering or concentrating

DA12

Difficulty walking or climbing steps

DA13

Difficulty washing all over or dressing

DA14A

Taken any benifit: ZAKAT & BAIT UL MALL

DA14B

Taken any benifit: BISP

DA14C

Taken any benifit: KHIDMAT CARD

DA14D

Taken any benifit: PENSION / RETIRMENT

DA14E

Taken any benifit: WATAN / HEALTH CARD

DA14X

Taken any benifit: OTHER

DA14Y

Taken any benifit: NOT ANY

DA14Z

Taken any benifit: DONT KNOW

DA14NR

Taken any benifit: NON RESPONSE

HH4

Supervisor number

HH5D

Day of interview

HH5M

Month of interview

HH5Y

Year of interview

HH6

Area

MLINE

Mother's line number

FLINE

Father's line number

helevel

Education of household head

disability

Functional difficulties

insurance

Health insurance

schage

Age at beginning of school year

melevel

Mother level of education

caretakerdis

Mother's functional disabilities (age 18-49 years)

felevel

Education of father

wscore

Combined wealth score

windex5

Wealth index quintile

windex10

Percentile Group of com1

wscoreu

Urban wealth score

windex5u

Urban wealth index quintile

windex10u

Percentile Group of urb1

wscorer

Rural wealth score

windex5r

Rural wealth index quintile

windex10r

Percentile Group of rur1

division

Division

majorcity

majorcity

hh6r

Area

suburban

hh7

Districts

hhweight

Household sample weight

stratum

stratum

psu

psu

References

UNICEF MICS (http://www.mics.unicef.org/surveys).

Examples

# Loading PakPMICS2018hl
 if (FALSE) {
 load(url(
  "https://github.com/myaseen208/PakPMICS2018Data/raw/master/PakPMICS2018hl.RData"
  ))
  PakPMICS2018hl
 }